Reorganize, add footer
