+++ src/app/loadout-drawer/LoadoutDrawer2.m.scss
@@ -1,0 +1,28 @@
+@import '../variables.scss';
+
+.body {
+  padding: 10px;
+}
+
+.notes {
+  margin-top: 4px;
+
+  summary {
+    font-weight: bold;
+    margin-bottom: 4px;
+    cursor: pointer;
+  }
+}
+
+// Group together buttons so they don't wrap onto different lines
+.inputGroup {
+  display: flex;
+  flex-flow: row nowrap;
+  align-items: center;
+  margin-right: 4px;
+  margin-bottom: 4px;
+  &:last-child {
+    margin-bottom: 0;
+  }
+  @include horizontal-space-children(4px);
+}
