+++ src/app/loadout-drawer/LoadoutDrawer2.m.scss.d.ts
@@ -1,8 +1,9 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'dimInput': string;
-  'loadoutName': string;
+  'body': string;
+  'inputGroup': string;
+  'notes': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
