+++ src/app/loadout-drawer/LoadoutDrawerDropTarget.tsx
@@ -22,9 +22,11 @@ export const bucketTypesSelector = createSelector(
 
 export default function LoadoutDrawerDropTarget({
   children,
+  className,
   onDroppedItem,
 }: {
   children?: React.ReactNode;
+  className?: string;
   onDroppedItem(item: DimItem): void;
 }) {
   const bucketTypes = useSelector(bucketTypesSelector);
@@ -41,7 +43,7 @@ export default function LoadoutDrawerDropTarget({
 
   return (
     <div
-      className={clsx({
+      className={clsx(className, {
         [styles.over]: isOver,
       })}
       ref={dropRef}
