+++ src/app/loadout-drawer/LoadoutDrawerFooter.m.scss.d.ts
@@ -1,8 +1,8 @@
 // This file is automatically generated.
 // Please do not change this file!
 interface CssExports {
-  'dimInput': string;
-  'loadoutName': string;
+  'inputGroup': string;
+  'loadoutOptions': string;
 }
 export const cssExports: CssExports;
 export default cssExports;
