+++ src/app/loadout-drawer/LoadoutDrawerHeader.m.scss
@@ -1,47 +1,16 @@
 @import '../variables.scss';
 
-.loadoutOptions {
-  width: 100%;
-
-  @include phone-portrait {
-    font-size: 14px;
-  }
-
-  form {
-    display: flex;
-    flex-flow: row wrap;
-  }
-
-  summary {
-    font-weight: bold;
-    margin-bottom: 4px;
-  }
-
-  details {
-    margin-top: 4px;
-  }
-}
-
-.secondary {
-  margin-left: auto;
-}
-
-// Group together buttons so they don't wrap onto different lines
-.inputGroup {
+.loadoutName {
+  font-size: 18px;
   display: flex;
   flex-flow: row nowrap;
   align-items: center;
-  margin-right: 4px;
-  margin-bottom: 4px;
-  &:last-child {
-    margin-bottom: 0;
-  }
-  @include horizontal-space-children(4px);
-}
+  max-width: 50em;
 
-.loadoutName {
-  width: 26em;
-  font-size: 18px;
+  svg,
+  :global(.app-icon) {
+    margin-right: 8px;
+  }
 }
 
 .dimInput {
@@ -55,7 +24,6 @@
   font-size: 18px;
 
   @include phone-portrait {
-    font-size: 14px;
     height: 32px;
   }
