+++ src/app/records/Records.m.scss
@@ -7,4 +7,7 @@
 
 .presentationNodeOptions {
   @include vertical-space-children(8px);
+  @include phone-portrait {
+    padding: 8px var(--inventory-column-padding) 0 var(--inventory-column-padding);
+  }
 }
